+++ src/app/child-dev-project/children/dashboard-widgets/children-count-dashboard/children-count-dashboard.component.scss
@@ -1,0 +1,1 @@
+@use '../../../../core/dashboard/dashboard-widget-base';
