+++ src/app/child-dev-project/children/dashboard-widgets/children-count-dashboard/children-count-dashboard.component.spec.ts
@@ -8,11 +8,11 @@ import {
 
 import { ChildrenCountDashboardComponent } from "./children-count-dashboard.component";
 import { MatCardModule } from "@angular/material/card";
-import { ChildrenService } from "../children.service";
+import { ChildrenService } from "../../children.service";
 import { RouterTestingModule } from "@angular/router/testing";
-import { Center, Child } from "../model/child";
+import { Center, Child } from "../../model/child";
 import { Observable } from "rxjs";
-import { ConfigurableEnumValue } from "../../../core/configurable-enum/configurable-enum.interface";
+import { ConfigurableEnumValue } from "../../../../core/configurable-enum/configurable-enum.interface";
 import { FontAwesomeTestingModule } from "@fortawesome/angular-fontawesome/testing";
 
 describe("ChildrenCountDashboardComponent", () => {
